test(github-client): lifecycleテストを別ファイルへ分離
【背景】test_github_client.pyの肥大化に伴い、責務別分割の一環としてlifecycle系テストを分離。
【変更内容】async context manager(__aenter__/__aexit__)とaclose()関連の12テスト関数をtest_github_client_lifecycle.pyへ純移動。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors
